Added functions for matrix operations
A simple and (so far) stable set of functions for linear algebra.
So far there are functions for creating and destroying matrixes;
load a matrix from file (plain text, row per line, space separated
values); print a matrix; set and get single elements; set a row.

Matrixes are implemented as structs with fields for number of
rows, columns, is_sqr, and data. The elements of the matrix are
stored in the data array of double type thus using a
linearly-storaged (row-major order) approach instead of a more
slower array of arrays.