Added: Makefile; embryonic main(); and .fdf files
All set for experimenting and testing!!!!

So far the Makefile DOES NOT compiles the MLX42 library which
was built manually.
The .gitignore file had to be temporarily edited to allow the
stagging of the linux version of MLX42: lib/MLX42/build/libmlx42.a.

At this point the .fdf files are correctly parsed and loaded as a
t_ft_mx matrix. Also the MLX42 library appears to be working: a
windows was created and a single pixel drawn :).