Added a new component 'w' for vertexes
In an unexpected turn of events, while writing what will be my
very flexible and scalable rendering function (during development
in fdf_start_gui()) I realized that the translation and scaling
matrixes (before this I was scaling and translating only my 2D
model by hand in fdf_scale_center()) require to be 4x4 matrixes.
After investigating a while it turned out that the practice is to
have a ONE-value 4th component for every vertex called 'w'.
BTW, may be this is the one that is eventually used for depth
(I was planning to eventually use the 'z' component for that).

Also the function ft_mx_mult() was fixed and the m3v 3vm functions
were discarded. A new ft_mx_mult3() function to multiply 3 matrixes
was added.