[Draft commit] Drawing edges with some bugs
I had to commit this version because it's a major success.
The edges are being drawn (with some bugs regarding the
non-orthogonal vertex neighbors).
The GUI responds to continous pressing of 'x' 'y' and 'z' which
rotates the model along the corresponding axis while re-rendering
and autofitting (auto-scale and auto-center).
This version has really messy code but I decided to commit it
represents a milestone.