Added a new matrix flip function
A ft_mx_flipc() function, analogous to existing ft_mx_flipr()
was added to allow flipping the map_mx matrix along the columns
axis (x axis). The intention is to provide tools to easily adjust
the source data to better align to the mathematics system of
coordinates and the pixel's image system.