Fixed fdf_get_vertex_mx() function
After rendering the edges it was detected a mismatch between the
expected order of vertexes columns in vertex_mx and the actual
order. It turned out that the i, j indexes in fdf_get_vertex_mx()
where interchanged. Also, because of this,  it was detected that
a non-square map_mx crashed the program.
This was fixed and now the edge rendering works as expected and
non-square maps are supported as the should.