Added a better handling of GUI events
Preparing for the remaining of the bonus part (user zooming and
translation) and in order to have a cleaner code and better files
structures, must of the code originally placed inside the GUI's
hooks was moved out to different functions and files. This events
filtering functions now call several 'feature' functions which
handle the actual work.

Dummy feature functions were written and tested from GUI's events,
they're ready to be used to implement the rest of the features.

A new logic was introduced through a new fdf member to signal that
re-rendering is pending.

Documentation was improved.