Object struct type created
For the tesselator an object type was created (along with its
functions to allocate and free). A member of fdf type was created
as a pointer to this new type. This type holds all the 3D object
data and is intended to move this data out from the fdf type.