FIX resize and funny pixels
The function to get minimum and maximum x and y values from the
camera space vertex matrix was generalized to be able to obtain
them from any vertex matrix. This was useful in troubleshooting
a rounding error in the function to obtain the autofit scale and
offset parameters that caused crashes when rendering points outside
the image boundaries after the latest changes in code. This also
caused some funny single pixels to be incorrectly drawn after
overflowing the width or height of the image.

At some point the re-rendering as response to a resize event stopped
working. This happened after the fdf_render() was optimized not to
destroy and re-create the mlx_image_t image every time it rendered.
Now this process was added to fdf_resizehook() after checking that
the new image size recalculation is equal or larger than 1x1 pixels.

The mechanism to request a re-render was changed to a new fdf_t member
which is first evaluated to see if current render request (from model to
no re-render) goes deeper than the new requirement, if not, the new
requirement is set. The new member render_request by levels replaced
the old true/false render_is_needed. The settings of
render_request were moved to the features functions (from the key down
events of rotation, translation and zoom). Now the fdf_render() function
is only called from the generic hook function and the rendering
is made according to the level set in fdf->render_request.

The resize event now triggers a call to a new feature function
(fdf_window_resize()) which takes care of the work in the same way
the other feature functions do. The fdf_recalc_wlayout() function
was moved as an auxiliary static function to the feature function
fdf_window_resize().

Makefile was improved by not fcleaning libft and by updating the list
of header files that should force a re-build.

The documentation was updated.