FIX missing pixels in high slope edges
The t_fdf_line type was modifyed to allow the drawing of a line
either as y = f(x) or x = g(y). The functions in fdf_drw_edges.c
where modified to draw lines with slopes greater than 1
as x = g(y), that is, iterating y values to obtain the correct x
values. In this way the rounding effect when passing from vertexes
double type values to integer type pixels do not cause missing
pixels when the lines are more vertical (higher slopes).