Functions added for vertex and surfaces math
Functions to create, substract and to calculate the cross product
of column vectors representing xyzw vertexes were added to the
lin_alg.h collection. These are needed to define a plane and
determine if a vertex or an edge belong to that plane. Also could
be useful if fragment rendering is ever implemented.