Added 3D rects related functions. Improved GUI.
A new type for the definition of a 3D rect was added to the
tesselator.h collection, as well as functions to create 3D rects
from two points (vertexes column vector matrixes); and to obtain
the Z value from an X or Y value. This is needed for definitively
solving the visibility problem of non-orthogonal edges.

fdf_is_in_plane() was renamed to fdf_point_is_in_plane() in
prevision for a possible function to evaluate if a rect is in
a plane.

The usability of the GUI was improved to ease the testing. The
user's keys to interact with the drawing are now more intuitive
and videogame-like.

More testing .fdf files added in data/.

Documentation was updated.