MAJOR UPDATE! Almost fully featured bonus version
REFACTORIZATION of edge visibility functions:
The function fdf_set_edge_visibility() used when tesselating
to determine those diagonal edges to be hidden, was refactored
into several auxiliary functions in fdf_set_edge_visibility.c.
For this, a new struct type to store four neighboring vertixes
and its create and destroy functions were added.

NEW alternate get_vertex functions for spherical and
cylindrical tesselating:
src/tesselator/fdf_get_vertex_mx_cyl.c
src/tesselator/fdf_get_vertex_mx_sph.c
Also an enum type for the tesselation type was added.
PENDING to change how the radius (sphere and cylinder) and
height (for the cylinder) is calculated. At this moment both
are hardcoded magic numbers.

3D object type now includes its source map_mx and its desired
tesselation type based on the corresponding enum type as new
struct members. The map_mx was simply a local variable to
fdf_main() function which has also changed.

The addition of the new members to object's type,  forced to
modify the fdf_create_object() to admit as argument
the source map_mx in order to create the 3D object which is
stored as one of the members of the t_fdf struct. Because of
this, the creation of the t_object was moved from
fdf_tesselate_map() to fdf_create_fdf(). That forced the
modification of fdf_tesselate_map() not to receive as an
argument the map_mx but the already created object that
contains it (points to); and not to return an object but rather
an int to signal errors. This will also ease the future
renderization of multiple objects: fdf_create_fdf() could
receive a list of maps and store a list of objects...
Consistently, the map tesselation (the calling of the
fdf_tesselate_map() function) was moved from fdf_main() to the
fdf_render() function as its most fundamental render request
level (the FROM_MODEL value of the tesselation enum type).

Finally, all of This allowed the addition of a new GUI
feature to toggle beetwen tesellation types as respond to
user's demand (plane to spherical, spherical to cylindrical,
and back to plane tesselation type).