FIX ft_mx_mult() ADDed ft_vx_dproduct()
I decided to re-write ft_mx_mult() due to its problems and poor
legibility. I decided to take an approach based on calculating
every element of the multiplication resulting matrix as the
dot product of the row-vector and col-vector. For this the
ft_vx_dproduct() version was added to the lin_alg collection.

Tests were made to assure no leaks and norminette compliance.

Also the fdf_main.c was commented with my conclusions regarding
the axis drawing convention and the ambiguity of the project's
subject on this regard.