ADDED ft_mx_fill function, WIP on ft_mx_load_file
A new ft_mx_fill() function was added to the lin_alg.h collection.

The work on ft_mx_load_file.c continues while maintaining
compatibility with the caller of ft_mx_load_file().
An array of two matrices to represent both the Z value and the
color value for each vertex was introduced to ft_mx_load_file().