WIP ft_mx_load_file.c, color matrix loaded
The work on ft_mx_load_file.c continues while maintaining
compatibility with the caller of ft_mx_load_file().
The matrix representing the map color value for each is succesfuly
filled with the color values.
The next steps will be to modify the ft_mx_load_file() caller
(the main()) and downstream from there.