WIP ft_mx_load_file.c, fdf_main.c
The work on fdf_main.c started, it now receives the array of
matrix pointers. For now, compatibility was maintained downstream.
The ft_mx_load_file() function was modified to allocate memory
for the array of matrix pointers since now has to return it to
main().
The next steps will be to modify fdf_create_fdf() and and
downstream from there.