WIP fdf_main.c fdf_get_vertex...
The work on the three fdf_get_vertex...() has started, they  now
receive the array of matrix pointers.
For now, compatibility was maintained downstream.
The next steps will be to modify the three fdf_get_vertex...()
functions and the vertex type so that the former retrieve color
information and storage it in the later.