Color support achieved
Now the line struct type has start and end color info.
Now when parsing the color info, the alpha channel is added as FF.
Now the default color if the map lacks color info is the full
RGBA value of 0xFFFFFFFF (opaque white).
An i_length member was added to line struct type to allow the fading
of colors.
The files fdf_get_color.c and fdf_calc_pixel_color.c were added with
color related functions.
A mechanism to force monochromatic rendering was added along with
the user feature that controls it.
The printing of instructions was updated and now prints after new
messages regarding the parsing of the file.
ONLY CLEANING UP IS LEFT before merging with development.